+++ src/columns/cgroup.rs
@@ -1,0 +1,45 @@
+use crate::process::ProcessInfo;
+use crate::{column_default, Column};
+use std::cmp;
+use std::collections::HashMap;
+
+pub struct Cgroup {
+    header: String,
+    unit: String,
+    fmt_contents: HashMap<i32, String>,
+    raw_contents: HashMap<i32, String>,
+    width: usize,
+}
+
+impl Cgroup {
+    pub fn new(header: Option<String>) -> Self {
+        let header = header.unwrap_or_else(|| String::from("Cgroup"));
+        let unit = String::new();
+        Cgroup {
+            fmt_contents: HashMap::new(),
+            raw_contents: HashMap::new(),
+            width: 0,
+            header,
+            unit,
+        }
+    }
+}
+
+#[cfg(any(target_os = "linux", target_os = "android"))]
+impl Column for Cgroup {
+    fn add(&mut self, proc: &ProcessInfo) {
+        let fmt_content = if let Ok(cgroups) = &proc.curr_proc.cgroups() {
+            cgroups
+                .last()
+                .map_or_else(|| "".to_string(), |x| x.pathname.to_string())
+        } else {
+            "".to_string()
+        };
+        let raw_content = fmt_content.clone();
+
+        self.fmt_contents.insert(proc.pid, fmt_content);
+        self.raw_contents.insert(proc.pid, raw_content);
+    }
+
+    column_default!(String);
+}