+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/dalance/procs/compare/v0.14.1...Unreleased) - ReleaseDate
 
+* [Added] Cgroup/Ccgroup column [#529](https://github.com/dalance/procs/issues/529)
+
 ## [v0.14.1](https://github.com/dalance/procs/compare/v0.14.0...v0.14.1) - 2023-10-06
 
 * [Added] Also look for a config file in /etc/procs/procs.toml [#533](https://github.com/dalance/procs/pull/533)